do not touch real files or directories when attempting to create backlinks/symlinks

+++ zocalo/wrapper/xia2.py
@@ -156,12 +156,18 @@ class Xia2Wrapper(dlstbx.zocalo.wrapper.BaseWrapper):
 
     if params.get('results_symlink'):
       path_elements = results_directory.split(os.sep)
+
+      # Full path to the symbolic link
       link_path = os.sep.join(path_elements[:-2] + [params['results_symlink']])
-      # because symlink can't be overwritten, create a temporary symlink in the child directory
-      # and then rename on top of potentially existing one in the parent directory.
-      tmp_link = os.sep.join(path_elements[:-1] + ['.tmp.' + params['results_symlink']])
-      os.symlink(os.sep.join(path_elements[-2:]), tmp_link)
-      os.rename(tmp_link, link_path)
+
+      # Only write symbolic link if a symbolic link is created or overwritten
+      # Do not overwrite real files, do not touch real directories
+      if not os.path.exists(link_path) or os.path.islink(link_path):
+        # because symlink can't be overwritten, create a temporary symlink in the child directory
+        # and then rename on top of potentially existing one in the parent directory.
+        tmp_link = os.sep.join(path_elements[:-1] + ['.tmp.' + params['results_symlink']])
+        os.symlink(os.sep.join(path_elements[-2:]), tmp_link)
+        os.rename(tmp_link, link_path)
 
     if not result['exitcode'] and not os.path.isfile('xia2.error') and os.path.exists('xia2.json'):
       self.send_results_to_ispyb()
